% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mixedBreak}
\alias{coef.mixedBreak}
\title{Extract mixed.break Coefficients}
\usage{
\usage{
\method{coef}{mixedBreak}(break.fit, type = "cluster")
}
}
\arguments{
\item{z}{an object of class \code{mixedBreak}}

\item{type}{[vector of character] CURRENTLY UNUSED. a character vector indicating 
which type of coefficients to extract and return. Several options include:
\itemize{
\item auc: area under the curve\n
\item breakpoint: position of the breakpoint(s)\n
\item breakpoint.range: Minimum and Maximum theoretical value of the
 breakpoint(s)\n
\item cluster: all subjects-specific regression coefficients (slopes + 
breakpoints) \n
\item corr: \n
\item duration: duration of each phases (same unit as segmented variables)\n
\item slope: slopes associated with each phases \n
\item R2: Nakagawa's coefficient of determination for mixed-model fits 
(fraction of variance explained by fixed-effects only)\n
\item R2_cond: Conditional coefficient of determination for mixed-model fits
(fraction of variance explained by fixed+random-effects)\n
\item pattern: pattern used
}}
}
\value{
a data.frame with requested coefficients. Eventually a list of 2 
for subjects-level & population-level statistics.
}
\description{
Extract summary statistics and regression coefficients from the mixed-effects
segmented model.
}
\examples{
data(SDIpsilo)
break.fit <- mixed.break(
  score ~ 0 + time + (0 + time|id),
  pattern = "10",
  data = SDIpsilo[SDIpsilo$time < 190, ]
)
coef(break.fit)
}
