% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.mixedBreak}
\alias{confint.mixedBreak}
\title{Confidence Intervals in segmented Mixed-Models}
\usage{
\method{confint}{mixedBreak}(z, level = 0.95)
}
\description{
Computes confidence intervals for all regression parameters, 
including the breakpoint, in a fitted ‘segmented mixed’ model.
}
