% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.break.R
\name{mixed.break}
\alias{mixed.break}
\title{Mixed-Effects Models with 1 or 2 breakpoint estimates}
\usage{
mixed.break(
  formula,
  data,
  pattern = NA,
  psi0 = NA,
  psi.range = NA,
  tol.logLik = 1e-04,
  it.max = 20L,
  approx = "Muggeo.LMM",
  psi.history = TRUE,
  x = FALSE,
  y = FALSE,
  dev.step = 0
)
}
\arguments{
\item{formula}{A two-sided linear formula object describing both the 
fixed-effects and random-effects part of the model, with the response on 
the left of `~` operator and the terms, separated by + operators, on the 
right. Random-effects terms are distinguished by vertical bars (|) 
separating expressions for design matrices from grouping factors.}

\item{data}{a data frame containing variables named in `formula`.}

\item{pattern}{[character] One of the 3 following options: \itemize{
\item '10' for a two-phases relationship with a non-null slope followed by
 a constant segment (second-phase slope is forced to 0) ;\n
\item '101' for a three-phases relationship with one non-null slope 
followed by a constant segment (second phase slope is forced to 0) and 
another non-null slope;\n
\item '1010' for a four-phases relationship with an alternating of linear 
non-null phases, and constant phases (phases 1 & 3 slopes are estimated, 
phases 2 & 4 slopes are null).
}}

\item{psi0}{[numeric] an optional starting value for the breakpoint 
estimates (first guess). 
It must be a numeric value in the range of the segmented variable. 
Default are quantile(s) of segmented variable with a response that is not 
\code{NA}. quantile are equally spaced in probability.}

\item{it.max}{[integer>0] Maximum number of iterations allowed. Default to
1\code{20L}. A warning is emitted is the maximum number of iterations is reached.}

\item{x, y}{[logical] If TRUE the corresponding components of the fit 
(the model matrix, the response) are returned.}

\item{lme.fit}{The fitted lme object at convergence.}

\item{fixed, random}{Fixed and Random-effects coefficients of the fit. G-like
variable (_i.e._ breakpoints) are on logit scale.}

\item{psi.i}{A matrix (n.ind x n.psi) containing the subject-specific breakpoints 
estimates (fixed + random).}

\item{fitted}{Vector of fitted values. It has the same length as the response 
in \code{data}.}
}
\value{
an list of class c('mixedBreak', 'mixedBreak{n.psi}') with n.psi being
the number of breakpoints specified. The most interesting component are:
}
\description{
This function fits a mixed-effects model with one or two
breakpoint estimates, in the formulation described in Muggeo (2014), 
allowing for random effects on the breakpoint estimates.
}
