% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.break1.R
\name{mixed.break1}
\alias{mixed.break1}
\title{Mixed-Effects Models with 1 breakpoint estimate}
\usage{
mixed.break1(
  formula,
  data,
  psi0 = NA,
  tol.logLik = 1e-04,
  it.max = 10L,
  pattern = "11",
  psi.history = TRUE,
  x = FALSE,
  y = FALSE,
  dev.step = 0
)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the 
fixed-effects and random-effects part of the model, with the response on 
the left of `~` operator and the terms, separated by + operators, on the 
right. Random-effects terms are distinguished by vertical bars (|) 
separating expressions for design matrices from grouping factors. 
By default, non-scalar random effects (where the design matrix has more 
than one column, e.g. (1+x|f)) are fitted with unstructured (general 
positive semidefinite) covariance matrices.}

\item{data}{a data frame containing variables named in `formula`.}

\item{psi0}{[numeric] an optional starting value for the breakpoint 
estimates (first guess). 
It must be a numeric value in the range of the segmented variable. 
Default is the middle of the observed segmented variable values 
(`mean(range(time))`).}

\item{it.max}{[integer>0] Maximum number of iterations allowed. Default to
10. A warning is emitted is the maximum number of iterations is reached.}

\item{pattern}{[character "11" | "10"] One of the two following options: \itemize{
\item '11' for two non-null slopes estimated ;\n",
\item '10' for one non-null slope followed by a constant segment"
}}

\item{x, y}{[logical] If TRUE the corresponding components of the fit 
(the model matrix, the response) are returned.}
}
\value{
TODO
}
\description{
This function fits a mixed-effects model with a single
breakpoint estimate, in the formulation described in Muggeo (2014), 
allowing for random effects on the breakpoint estimate.
}
