% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseTraj.R
\name{noiseTraj}
\alias{noiseTraj}
\title{Simulate "10"-alike trajectories from known parameters and noise level}
\usage{
noiseTraj(
  true.traj,
  n.obs,
  score.sd,
  times.sd,
  breakpoints.sd = NULL,
  break.x.dist = 0,
  outlier.prob = 0,
  na.prob = 0,
  n.trail = 0L
)
}
\arguments{
\item{n.obs}{[integer >= 1] number of observations to simulate in the dataset.}

\item{score.sd}{[numeric>0] noise level (sd) on the measurement/score.
TODO - later could be a vector to specify noise level in each phase of the experiment
(typically lower during plateau phase)}

\item{times.sd}{[numeric>0]}

\item{breakpoints.sd}{[data.frame] data.frame with \code{length(pattern)+1} rows
and 2 columns: \itemize{
\item "bp.x.sd" noise (sd) parameter for breakpoints' x coordinates.
\item "bp.y.sd" noise (sd) parameter for breakpoints' y coordinates.
If \code{pattern[i]==0}, there should be \code{bp.y.sd[i]==0}.
}
In case of conflict between pattern and breakpoints specifications, the pattern prevails
and the previous breakpoint specification will override the next one.
A warning message will be displayed.}

\item{break.x.dist}{[numeric>0] Minimum between-breakpoints distance on
x coordinates. Default is 0.}

\item{outlier.prob}{[0<numeric<1] proportion of outlier in the dataset.
Outliers are simulated as a random drop of 1-3 points in SDI, with higher probability
on}

\item{na.prob}{[0<numeric<1] proportion of missing at random values in the dataset.}

\item{n.trail}{[integer] number of trailing observations at the end of the trajectory.
Default is 0. Has a meaning when positive. 
If negative, it will remove the last observations from the trajectory.
Ignored if pattern ends with a 0 / plateau.}
}
\value{
A list of 2 datasets: \itemize{
\item "sim.dataset" contains ID, time, true, noised measurements and score,
as well as outlier / missing / trailing flags
\item "sim.gen.model" contains ID and patients-specific breakpoints coordinates
}
}
\description{
Simulates trajectories of SDI in patients after psilocybin intake from
a dataset that contains the true values of parameters. 
Noise level can be specified for every parameter specified in the reference dataset.
}
