% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulate SDI trajectories with known truth and noise}
\usage{
simData(
  n.obs,
  score.sd,
  times = list(value = 20, sd = 0),
  plateau = list(value = 9, value.sd = 3, duration = 82.4, duration.sd = 33.1),
  b.onset = NULL,
  b.return = NULL,
  breakpoints = NULL,
  ending.times = NULL,
  outlier.prob = 0,
  na.prob = 0,
  n.trail = 0L
)
}
\arguments{
\item{n.obs}{[integer >= 1] number of observations to simulate in the dataset.}

\item{score.sd}{[numeric>0] noise level (sd) on the measurement.
TODO - later could be a vector to specify noise level in each phase of the experiment
(typically lower during the plateau phase)}

\item{times}{[named list or integer vector] times and noise (if any) of simulated measurements,
for regular strategies.
Can be a vector with values of time measurements to specify other irregular strategies.}

\item{plateau}{[named list] named list with \itemize{
\item "value" for the height / altitude of the plateau (e.g. its 'y' coordinate).
\item "value.sd" for the noise level of height of plateau.
It is *highly recommended* to specify plateau's height given its constraint (\eqn{\leq 10}),
to avoid any unintended re-computation of \code{b.onset} or \code{break.1}. 
They would then not match the pre-specified distribution for those parameters.
\item "duration" for the duration of the plateau on the same time scale as \code{times}.
If \code{plateau$duration==0}, then it is assumed that there is no plateau, and the 
second breakpoint value is ignored if provided.
\item "duration.sd" for the noise level of duration of plateau (on the same time scale 
as \code{times}).
\item "duration.min" Minimum duration of simulated plateau, if any (always >= 0)
\item "duration.max" Maximum duration of simulated plateau, if any
}}

\item{b.onset}{[named list] named list with \itemize{
\item "value" for coefficient value,
\item "sd" for the noise level,
\item "min" for minimum value simulated.
}}

\item{b.return}{[named list] named list with \itemize{
\item "value" for coefficient value,
\item "sd" for the noise level,
\item "max" for maximum value simulated.
}}

\item{breakpoints}{[data.frame] data.frame with \itemize{
\item "bp" first column containing the two breakpoints (sorted,
should be between 0 and the end of the trajectory)
\item "bp.sd" second column with noise parameter for each breakpoint.
}}

\item{ending.times}{[named list] named list with \itemize{
\item "value" for average ending time of effects on the patients
\item "sd" for the noise level of parameter
\item "max" for maximum value simulated.
}}

\item{outlier.prob}{[0<numeric<1] proportion of outlier in the dataset.
Outliers are simulated as a random drop of 1-3 points in SDI.}

\item{na.prob}{[0<numeric<1] proportion of missing at random values in the dataset.}

\item{n.trail}{[integer >= 1] number of trailing observations at the end of the trajectory}
}
\value{
A dataset with n.obs trajectories of SDI, simulated according to specified parameters
It contains ID, time of measurements, Data Generation Model, simulated measurements
and other characteristics such as outlier or missing value flag
}
\description{
Simulates trajectories of SDI in patients after psilocybin intake.
The simulated trajectories have 3 separate linear phases: onset, peak/plateau,
return to normal consciousness. There are many different parametric representations of such
trajectories. In case it is over specified, *the first parameters will prevail*.
The SDI scale is an integer subjective scale ranging from 0-10.
}
