% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData10.R
\name{simData10}
\alias{simData10}
\title{Simulate "10"-alike trajectories with known parameters}
\usage{
simData10(
  n.obs,
  score.sd,
  times = list(value = 20, sd = 0),
  breakpoints = NULL,
  break.min.dist = list(x = 0, y = NA),
  outlier.prob = 0,
  na.prob = 0,
  n.trail = 0L
)
}
\arguments{
\item{n.obs}{[integer >= 1] number of observations to simulate in the dataset.}

\item{score.sd}{[numeric>0] noise level (sd) on the measurement.
TODO - later could be a vector to specify noise level in each phase of the experiment
(typically lower during plateau phase)}

\item{times}{[named list or integer vector] times and noise (if any) of simulated measurements,
for regular strategies.
Can be a vector with values of time measurements to specify other irregular strategies.}

\item{breakpoints}{[data.frame] data.frame with \code{length(pattern)+1} rows
and 5 columns: \itemize{
\item "pattern" shape of the pattern. Should only contain 1s or 0s,
never two 0 in a row. The last value is ignored and set to \code{NA}.
The pattern of the trajectory has as many segments as \code{length(pattern)},
and 1 breakpoint more than the length of pattern (beginning and ending are considered breakpoints).
Specifying 1 means a slope coefficients is expected to be non-zero, and 0 means a plateau (slope=0).
\item "bp.x" every breakpoints x coordinates on the time scale (sorted,
should be between 0 and the end of the trajectory).
\item "bp.y" every breakpoints y coordinates on the SDI scale (between 0 and 10).
If \code{pattern[i]==0}, then there should be \code{bp.y[i-1] == bp.y[i]}.
\item "bp.x.sd" noise (sd) parameter for breakpoints' x coordinates.
\item "bp.y.sd" noise (sd) parameter for breakpoints' y coordinates.
If \code{pattern[i]==0}, there should be \code{bp.y.sd[i]==0}
}
In case of conflict between pattern and breakpoints specifications, the pattern prevails
and the previous breakpoint specification will override the next one.
A warning message will be displayed.}

\item{break.min.dist}{[list of numeric] Minimum between-breakpoints distance on
x & y coordinates. A list with 2 items: \itemize{
\item "x" [numeric>0] for x lower bound on distance. Default is 0.
\item "y" *IGNORED* [numeric>0] for y lower bound on distance. Default is NA.
}}

\item{outlier.prob}{[0<numeric<1] proportion of outlier in the dataset.
Outliers are simulated as a random drop of 1-3 points in SDI.}

\item{na.prob}{[0<numeric<1] proportion of missing at random values in the dataset.}

\item{n.trail}{[integer] number of trailing observations at the end of the trajectory.
Default is 0. Has a meaning when positive. 
If negative, it will remove the last observations from the trajectory.
Ignored if pattern ends with a 0 / plateau}
}
\value{
A dataset with n.obs trajectories of SDI, simulated according to specified parameters
It contains ID, time of measurements, Data Generation Model, simulated measurements
and other characteristics such as outlier or missing value flag
}
\description{
Simulates trajectories of SDI in patients after psilocybin intake.
Any "10"-alike pattern can be specified with breakpoints coordinates and noise level.
The SDI scale is an integer subjective scale ranging from 0-10.
}
