% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mixedBreak}
\alias{summary.mixedBreak}
\title{Summarize model fits for segmented mixed-effects regression}
\usage{
\method{summary}{mixedBreak}(z, default = FALSE)
}
\arguments{
\item{z}{an object of class "mixedBreak", usually, a result of a call to
`mixed.break` (TODO - make link to mixed.break documentation)}

\item{default}{[logical] use default list summary (default is FALSE).}
}
\value{
Returns the fitted object \code{z}.
}
\description{
\code{summary} method for class \code{mixed.break}
}
\details{
This function computes and print a list of summary statistics of the
fitted segmented linear mixed-effects model
}
